!refactor: move koala colours and utils to new implementation
