+++ koala/db.py
@@ -18,7 +18,7 @@ from sqlalchemy.orm import sessionmaker
 # Own modules
 from koala.env import DB_KEY, ENCRYPTED_DB
 from koala.models import mapper_registry, KoalaExtensions, GuildExtensions
-from koala.utils.KoalaUtils import get_arg_config_path, format_config_path
+from koala.utils import get_arg_config_path, format_config_path
 
 
 # Constants
